HS-80 feat: 루틴 삭제

+++ src/main/java/com/holeinone/ssafit/model/service/VideoService.java
@@ -26,5 +26,5 @@ public interface VideoService {
     YoutubeVideo directYoutubeUrl(String url, String part, int sequence ,int restSecondsAfter) throws CustomException;
 
     //루틴 아이디를 통해 루틴 삭제
-    int routineIdDelete(int routineId);
+    int routineIdDelete(long routineId);
 }
